requirements.txt
streamlit==1.55.0
google-cloud-bigquery==3.40.1
google-auth==2.49.1
db-dtypes==1.5.0
pandas==2.3.3
